se mejora login y pagina de registro, se comienza pagina de recepciones

+++ lib/models/provedor.dart
@@ -1,7 +1,7 @@
 import 'dart:ffi';
 
 class Provedor {
-   Long idProv;
+   int idProv;
 	
 	 String nombreProv;
